various tweaks
