+++ cmd/vvgo-uploader/main.go
@@ -1,0 +1,5 @@
+package main
+
+func main() {
+
+}
