+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/go-redis/redis/v7 v7.2.0
 	github.com/minio/minio-go/v6 v6.0.50
 	github.com/sirupsen/logrus v1.4.2
+	github.com/stretchr/testify v1.2.2
 	golang.org/x/net v0.0.0-20190923162816-aa69164e4478
 	golang.org/x/sys v0.0.0-20200331124033-c3d80250170d // indirect
 )
