--- pkg/storage/object_test.go
@@ -1,1 +1,0 @@
-package storage
